Example of pixelized 3D view combined with non-pixelated GUI.
Main view is now world2d which contains the HUD w/ control help bubble plus TextureRect + SubViewport + instance-of world3d.

The Subviewport is configured to render at about 1/4 of the default project screen res.

With this division it's also possible to just play the world3d scene directly (in the editor) to see it unpixelated.